fix: export metadata only in server environment
